refactor(nip19): move inner classes
