+++ nostr/nostr-nip19/src/main/java/org/tbk/nostr/nip19/Codec.java
@@ -1,0 +1,7 @@
+package org.tbk.nostr.nip19;
+
+interface Codec<T> {
+    boolean supports(String hrp, Class<?> clazz);
+
+    T decode(String hrp, byte[] data);
+}
