+++ nostr/nostr-nip19/src/main/java/org/tbk/nostr/nip19/EntityType.java
@@ -1,0 +1,20 @@
+package org.tbk.nostr.nip19;
+
+import lombok.Getter;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+enum EntityType {
+    NPUB("npub"),
+    NSEC("nsec"),
+    NOTE("note"),
+    NPROFILE("nprofile"),
+    NEVENT("nevent"),
+    NADDR("naddr"),
+    ;
+
+    @NonNull
+    private final String hrp;
+}
